requirements.txt
alembic==1.17.2
amqp==5.3.1
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
billiard==4.2.3
blinker==1.9.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.0.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
fpdf==1.7.2
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
paramiko==4.0.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gvm==26.7.1
python-multipart==0.0.20
python-nmap==0.7.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
rich==14.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==2.0.1
